Module spandex::layout::paragraphs::utils::linebreak

Utility functions for the various stages of the line breaking algorithm.

Structs

Measures

Accumulator to hold the three key related measures.

Functions

compute_adjustment_ratio

Computes the adjusment ratio of a line of items, based on their combined width, stretchability and shrinkability. This essentially tells how much effort has to be produce to fit the line to the desired width.

compute_adjustment_ratios_with_breakpoints

Computes the adjustment ratios of all lines given a set of line lengths and breakpoint indices. This allows to speed up the adaptation of glue items.

compute_demerits

Computes the demerits of a line based on its accumulated penalty and badness.

compute_fitness

Computes the fitness class of a line based on its adjustment ratio.

create_node_for_feasible_breakpoint

Handles a feasible breakpoint and adds it to the current graph of feasible breakpoints if it’s good enough.

find_legal_breakpoints

Finds all the legal breakpoints within a paragraph. A legal breakpoint is an item index such that this item is either a peanalty which isn’t infinite or a glue following a bounding box.

get_measures_to_next_box

Computes the accumulated measures from the current linebreak to the next bounding box in the provided items.

is_forced_break

Checks whether or not a given item encodes a forced linebreak.