Module spandex::parser

This crate contains the parser for spandex.

Re-exports

pub use combinators::parse;

Modules

ast
This module contains everything related to the ast.

combinators
This module contains all the functions needed for parsing.

error
This module contains everything related to parsing errors.

utils
This module contains some functions that will help us managing strings.

warning
This module contains everything related to parse warnings.

Structs

Parsed
An ast that was successfully parsed.

Position
A position is a span but without the reference to the complete str.

Functions

position
Returns the position of a span.

Type Definitions

Span
This type will allow us to know where we are while we're parsing the content.