Module spandex::layout::paragraphs

Logic for laying out a paragraph.

Modules

engine
Mathematical logic for typesetting a sequence of words which have a semantics of "paragraph". That is, the logic to split a sequence of words into lines.

graph
Structure used in the graph part of the line breaking algorithm. It allows to store the accumulated statistics of line breaks and allows a dynamic programming approach.

items
Various blocks holding information and specifications about the structure of a paragraph.

justification
This module contains the trait and implementation of justification algorithms.

ligatures
This module contains the functions related to ligatures.

utils
Set of utility functions for manipulating ASTs, the line breaking algorithm, and the Paragraph structure.

Structs

Paragraph
Holds a list of items describing a paragraph.