Module spandex::parser::combinators

This module contains all the functions needed for parsing.

Functions

error
Creates an error.

get_block
Gets a block of content.

parse
Parses a whole dex file from a name.

parse_any
Parses some multiline inline content.

parse_block_content
Parses a block of content.

parse_bold
Parses some bold content.

parse_comment
Parses a comment.

parse_content
Parses a whole dex file.

parse_delimited
Parses a delimited element.

parse_group
Parses some text content.

parse_inline_math
Parses some math inline math.

parse_italic
Parses some italic content.

parse_paragraph
Parses a paragraph of text content.

parse_single_line
Parses a title on a single line.

parse_title
Parses a whole title.

parse_title_level
Parses the hashes from the level of a title.

should_stop
Returns true if the character passed as parameter changes the type of parsing we're going to do.

warning
Creates a warning.